requirements.txt
asgiref==3.5.2
crispy-tailwind==0.5.0
Django==4.0.6
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-environ==0.9.0
djangorestframework==3.13.1
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
